Functional codes to simulate and reconstruct
The current push has all the mains required to simulate and reconstruct the holograms systematically. The focos_main file allows for an easy parameter swap to focus all the holograms using the three methods. Next steps will involve the modification of the FOV in kreuzer to quantify the distortion accordingly